large-fullstack/requirements.txt
flask
Flask-SQLAlchemy
psycopg2-binary
redis
pika